Falta agregar los precios

+++ templates/register.py
@@ -1,32 +1,46 @@
-# import tkinter
 from tkinter import *
-from templates.shop import shopWindow
 from tkinter import messagebox
+import os.path
+from templates.shop import shopWindow
 
 
 def saveUser():
+    # Obtenemos el nombre de usuario y la contraseña
     username = user_entry.get().strip()
     userPassword = pass_entry.get()
+    userDni = dni_entry.get().strip()
 
+    # Creamos un archivo con el nombre del usuario
     routeFile = f"./src/users/{username}.txt"
-    # Guardamos la contraseña en el archivo
-    with open(routeFile, "w") as file:
-        if len(userPassword) > 3:
-            file.write(userPassword)
-            file.close()
 
-            # Re dirigimos a la ventana de Ventas
-            root.destroy()
-            shopWindow()
-        else:
-            messagebox.showwarning(
-                'Login Page', 'La contraseña debe tener al menos 4 caracteres')
+    # Existe el archivo?
+    existingUser = os.path.isfile(routeFile)
+
+    # Si retorna True es porque el usuario ya existe
+    if existingUser:
+        messagebox.showwarning(
+            'Registro', 'Usuario ya registrado')
+
+    else:
+        with open(routeFile, "w") as file:
+            if len(userPassword) > 3:
+                # Guardamos la contraseña en el archivo
+                file.write(f"Constraseña : {userPassword} | DNI : {userDni}")
+                file.close()
+
+                # Re dirigimos a la ventana de Ventas
+                root.destroy()
+                shopWindow()
+            else:
+                messagebox.showwarning(
+                    'Registro', 'La contraseña debe tener al menos 4 caracteres')
 
 
 def RegisterWindow():
     global root
     global user_entry
     global pass_entry
+    global dni_entry
     root = Tk(className='Login')
 
     # Agregamos un background image
@@ -41,17 +55,21 @@ def RegisterWindow():
     # Entry username
     user_entry = Entry(root, font=('Helvetica', 14, 'normal'))
     user_entry.focus_set()
-    user_entry.place(x=160, y=248, width=200, height=25)
+    user_entry.place(x=170, y=228, width=200, height=25)
+
+    # Entry DNI
+    dni_entry = Entry(root, font=('Helvetica', 14, 'normal'))
+    dni_entry.place(x=170, y=312, width=200, height=25)
 
     # Entry password
     pass_entry = Entry(root, font=('Helvetica', 14, 'normal'), show="*")
-    pass_entry.place(x=160, y=345, width=200, height=25)
+    pass_entry.place(x=170, y=396, width=200, height=25)
 
     # Button Register
     buttonRegister = Button(root, text='Registrarse', font=('Helvetica', 14, 'normal'),
                             bg="black", fg="white", command=saveUser)
 
-    buttonRegister.place(x=195, y=428, width=220, height=36)
+    buttonRegister.place(x=210, y=460, width=220, height=36)
 
     root.mainloop()
